Rename init and host files
